# Fix clean network example performance cf