--- a/NEWS.md
+++ b/NEWS.md
@@ -5,7 +5,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # [Unreleased]
 
-# Version 0.1.0
+# Version 0.1.1 - 2025-06-24
+
+## Fixed
+
+- Examples that take too long to run are only run in interactive mode
+
+# Version 0.1.0 - 2025-06-23
 
 ## Added