--- a/cran-comments.md
+++ b/cran-comments.md
@@ -3,3 +3,10 @@
 0 errors | 0 warnings | 1 note
 
 * This is a new release.
+
+## Resubmission
+
+This is a resubmission. In this version:
+
+* The example of `clean_network()` flagged as taking too long
+is run only in interactive mode.