--- a/man/clean_network.Rd
+++ b/man/clean_network.Rd
@@ -22,9 +22,11 @@ Subdivide edges by \href{https://luukvdmeer.github.io/sfnetworks/articles/sfn02_
 and discard all but the main connected component.
 }
 \examples{
+\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 bucharest_osm <- get_osm_example_data()
 edges <- dplyr::bind_rows(bucharest_osm$streets,
                           bucharest_osm$railways)
 network <- sfnetworks::as_sfnetwork(edges, directed = FALSE)
 clean_network(network)
+\dontshow{\}) # examplesIf}
 }