@cforgaci 2025-06-24
> Thanks, was `valley.R` accidentally left out from the commit?

No, that example was updated before in `valley.R`; only the documentation was not.