@cforgaci 2025-06-24
> The CRAN submission flagged the example in `clean_network()` as taking too long in “Examples with CPU (user + system) or elapsed time > 5s” or "> 10s". See logs below. With this PR, the example is only run in intercative mode. I also updated the version to 0.1.1.

@fnattino, thanks, I missed that. Updated now.